BE/External/requirements.txt
flask
joblib
flask-cors
scikit-learn==1.2.2
numpy
pandas
python-dotenv
requests